разбивает код по модулям
